Phase 2 commit 2: extract domain/yaml.ts
The gray-matter config (custom YAML engine that preserves Golarian
date strings) and the parse/serialise/extractTitle helpers are pure
domain logic — no IO, no protocol awareness. Move them to
domain/yaml.ts so adapters and other domain modules can share the
single sanctioned parse/serialise path.

api.ts loses its direct gray-matter and js-yaml imports; future
domain extractions go through this module.

No behaviour change. Server tests (33) green.
